disable unsubscribe northwestern events

+++ src/pages/item-details/item-details.ts
@@ -2,6 +2,7 @@ import { Component, Input } from '@angular/core';
 import { NavController, NavParams } from 'ionic-angular';
 import { UserService } from '../../services/user-service';
 import { Event } from '../../models/event'
+import { AlertController } from 'ionic-angular';
 import * as moment from 'moment';
 import * as firebase from "firebase";
 
@@ -18,7 +19,7 @@ export class ItemDetailsPage {
   time: string;
   priorView: string;
 
-  constructor(public navCtrl: NavController, public navParams: NavParams, private userService: UserService) {
+  constructor(public navCtrl: NavController, public alertCtrl: AlertController, public navParams: NavParams, private userService: UserService) {
     // If we navigated to this page, we will have an item available as a nav param
     this.selectedItem = navParams.get('item');
     this.priorView = navParams.get('view');
@@ -41,6 +42,15 @@ export class ItemDetailsPage {
     // console.log(JSON.stringify(this.selectedItem.calendartype));
     var user = firebase.auth().currentUser;
     var uid = user.uid;
+    if(this.selectedItem.calendartype=='northwestern'){
+      let alert = this.alertCtrl.create({
+        title: 'Alert!',
+        subTitle: 'You cannot unsubscribe Northwestern events since you are a student in Northwestern.',
+        buttons: ['OK']
+      });
+      alert.present();
+      return;
+    }
     console.log(uid);
     if (this.priorView == 'PersonalPage'){
        this.userService.removeUserSubscriptions(uid, this.selectedItem.calendartype);
@@ -50,7 +60,7 @@ export class ItemDetailsPage {
       this.userService.updateUserSubscriptions(uid, this.selectedItem.calendartype);
       console.log('hello');
     }
-    
+
   }
 
   displaybuttonname(){
